requirements.txt
Flask==3.1.1
python-dotenv==1.0.0
bcrypt==4.0.1
werkzeug==2.3.4
Jinja2==3.1.2
cloudinary==1.39.0